minor adjustments to upgrades
tried adding start-percentage="100" but that doesn't have an effect on
newly-produced units; they are produced with the hp specified in their
unit file, and it increases to max hp using the regeneration rate.

mystical forest: Removed forest guardian from affected units, because
this is required to build them, so better to just give the guard the
attributes specified here instead.

removed forest guardian from elven lore, because again, can't be built
yet.